17-ый commit исправление замечаний
