+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -1,53 +1,61 @@
 package com.example.playlistmaker
 
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-
 data class Track(
     val trackName: String,
     val artistName: String,
     val trackTime: String,
     val artworkUrl100: String
-)
-
-class TrackHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-    private val trackName: TextView = itemView.findViewById(R.id.trackNameTextView)
-    private val artistName: TextView = itemView.findViewById(R.id.artistNameTextView)
-    private val duration: TextView = itemView.findViewById(R.id.durationTextView)
-    private val cover: ImageView = itemView.findViewById(R.id.coverAlbum)
-
-    fun bind(model: Track) {
-        trackName.text = model.trackName
-        artistName.text = model.artistName
-        duration.text = model.trackTime
-        Glide.with(itemView)
-            .load(model.artworkUrl100)
-            .placeholder(R.drawable.place_holder_img)
-            .transform(RoundedCorners(10))
-            .centerCrop()
-            .into(cover)
+) {
+    companion object {
+        fun testTrackList(): MutableList<Track> {
+            val trackList: MutableList<Track> = mutableListOf()
+            trackList.add(
+                Track(
+                    "Smells Like Teen Spirit",
+                    "Nirvana",
+                    "5:01",
+                    "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
+                )
+            )
+            trackList.add(
+                Track(
+                    "Billie Jean",
+                    "Michael Jackson",
+                    "4:35",
+                    "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
+                )
+            )
+            trackList.add(
+                Track(
+                    "Stayin' Alive",
+                    "Bee Gees",
+                    "4:10",
+                    "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
+                )
+            )
+            trackList.add(
+                Track(
+                    "Whole Lotta Love",
+                    "Led Zeppelin",
+                    "5:33",
+                    "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
+                )
+            )
+            trackList.add(
+                Track(
+                    "Sweet Child O'Mine",
+                    "Guns N' Roses",
+                    "5:03",
+                    "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
+                )
+            )
+            return trackList
+        }
     }
 
 }
 
-class TrackAdapter(private val tracks: List<Track>) : RecyclerView.Adapter<TrackHolder>() {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.track_list, parent, false)
-        return TrackHolder(view)
-    }
 
-    override fun onBindViewHolder(holder: TrackHolder, position: Int) {
-        holder.bind(tracks[position])
-    }
 
-    override fun getItemCount(): Int {
-        return tracks.size
-    }
-}
+
+
