+++ app/src/main/java/com/example/playlistmaker/TrackAdapter.kt
@@ -1,0 +1,20 @@
+package com.example.playlistmaker
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+
+class TrackAdapter(private val tracks: List<Track>) : RecyclerView.Adapter<TrackHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.track_list, parent, false)
+        return TrackHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: TrackHolder, position: Int) {
+        holder.bind(tracks[position])
+    }
+
+    override fun getItemCount(): Int {
+        return tracks.size
+    }
+}
