+++ app/src/main/res/layout/track_list.xml
@@ -13,7 +13,7 @@
         android:layout_marginVertical="@dimen/margin_8dp"
         android:layout_marginStart="@dimen/margin_13dp"
         android:layout_marginEnd="@dimen/margin_8dp"
-        app:srcCompat="@drawable/place_holder_img" />
+        app:srcCompat="@drawable/placeholder_ico" />
 
     <LinearLayout
         android:id="@+id/trackInfo"
@@ -31,14 +31,15 @@
             tools:text="Большое и длинное название песни" />
 
         <LinearLayout
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:orientation="horizontal">
 
             <TextView
                 android:id="@+id/artistNameTextView"
-                android:layout_width="wrap_content"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
+                android:layout_weight="1"
                 style="@style/TrackInfoTextView"
                 app:drawableEndCompat="@drawable/tactical_dot_ico"
                 tools:text="Исполнитель" />
