+++ app/src/main/res/values/dimens.xml
@@ -11,6 +11,8 @@
     <dimen name="head_text_size">26sp</dimen>
     <dimen name="regular_text_size">19sp</dimen>
     <dimen name="small_text_size">13sp</dimen>
+    <dimen name="text_size_11sp">11sp</dimen>
+    <dimen name="text_size_16sp">16sp</dimen>
     <dimen name="search_corner_radius">8dp</dimen>
     <dimen name="drawable_padding">16dp</dimen>
     <dimen name="search_text_size">16dp</dimen>
