+++ app/src/main/res/values/styles.xml
@@ -86,7 +86,7 @@
     </style>
 
     <style name="TrackNameTextView">
-        <item name="android:textSize">@dimen/regular_text_size</item>
+        <item name="android:textSize">@dimen/text_size_16sp</item>
         <item name="android:textColor">?attr/colorOnPrimary</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
         <item name="android:ellipsize">end</item>
@@ -94,7 +94,7 @@
     </style>
 
     <style name="TrackInfoTextView">
-        <item name="android:textSize">@dimen/small_text_size</item>
+        <item name="android:textSize">@dimen/text_size_11sp</item>
         <item name="android:textColor">@color/Track_List_Under_Name_Color</item>
         <item name="android:fontFamily">@font/ys_display_regular</item>
         <item name="iconTint">@color/Track_List_Under_Name_Color</item>
